app/requirements.txt
streamlit
pandas
numpy
matplotlib
altair
requests
python-dotenv
streamlit-aggrid
azure-identity>=1.12.0
azure-keyvault-secrets>=4.7.0
msal>=1.24.0


